# [PERF] stock: improve _process_decrease with serial numbers