Description
On a 'stock.move' with a lot of move lines (serial numbers), setting the 'stock.move' quantity to 0 could take a few minutes due to the number of unlink calls.

Solution:
Extract unlink call outside the "for" loop.

| Move Lines | Without Fix | With Fix |
| --- | --- | ---|
| 10 | 321 ms | 102 ms |
| 100 | 2.60 s | 108 ms |
| 1.000 | 31.71 s | 380 ms |
| 10.000 | 20.7 min | 1.41 s |
| 100.000 | no idea | 17.47 |

---

OPW-4043076




---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr